Crate jsonrpsee_ws_server

jsonrpsee-ws-server

jsonrpsee-ws-server is a JSON RPC WebSocket server library that’s is built for async/await.

Re-exports

pub use jsonrpsee_types as types;

pub use tracing;

Structs

NoopIdProvider

No-op implementation to be used for servers that don’t support subscriptions.

PendingSubscription

Represent a pending subscription which waits until it’s either accepted or rejected.

RandomIntegerIdProvider

Generates random integers as subscription ID.

RandomStringIdProvider

Generates random strings of length len as subscription ID.

RpcModule

Sets of JSON-RPC methods can be organized into a “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

SubscriptionSink

Represents a single subscription.

WsServer

A WebSocket JSON RPC server.

WsServerBuilder

Builder to configure and create a JSON-RPC Websocket server

WsServerHandle

Handle that is able to stop the running server or wait for it to finish its execution.

WsShutdownWaiter

A Future that resolves once the server has stopped.

Traits

IdProvider

Trait to generate subscription IDs.