Struct jsonrpsee_ws_server::WsServerBuilder

pub struct WsServerBuilder<M = ()> { /* private fields */ }

Builder to configure and create a JSON-RPC Websocket server

Implementations

impl Builder

pub fn new() -> Self

Create a default server builder.

impl<M> Builder<M>

pub fn max_request_body_size(self, size: u32) -> Self

Set the maximum size of a request body in bytes. Default is 10 MiB.

pub fn max_response_body_size(self, size: u32) -> Self

Set the maximum size of a response body in bytes. Default is 10 MiB.

pub fn max_connections(self, max: u64) -> Self

Set the maximum number of connections allowed. Default is 100.

pub fn batch_requests_supported(self, supported: bool) -> Self

Enables or disables support of batch requests. By default, support is enabled.

pub fn max_subscriptions_per_connection(self, max: u32) -> Self

Set the maximum number of connections allowed. Default is 1024.

pub fn register_resource(
self,
label: &'static str,
capacity: u16,
default: u16
) -> Result<Self, Error>

Register a new resource kind. Errors if label is already registered, or if the number of registered resources on this server instance would exceed 8.

See the module documentation for resurce_limiting for details.

pub fn set_middleware<T: Middleware>(self, middleware: T) -> Builder<T>

Add a middleware to the builder Middleware.

use std::time::Instant;

use jsonrpsee_core::middleware::Middleware;
use jsonrpsee_ws_server::WsServerBuilder;

#[derive(Clone)]
struct MyMiddleware;

impl Middleware for MyMiddleware {
    type Instant = Instant;

    fn on_request(&self) -> Instant {
        Instant::now()
    }

    fn on_result(&self, name: &str, success: bool, started_at: Instant) {
        println!("Call to '{}' took {:?}", name, started_at.elapsed());
    }
}

let builder = WsServerBuilder::new().set_middleware(MyMiddleware);

pub fn custom_tokio_runtime(self, rt: Handle) -> Self

Configure a custom tokio::runtime::Handle to run the server on.

Default: tokio::spawn

pub fn ping_interval(self, interval: Duration) -> Self

Configure the interval at which pings are submitted.

This option is used to keep the connection alive, and is just submitting Ping frames, without making any assumptions about when a Pong frame should be received.

Default: 60 seconds.

Examples

use std::time::Duration;
use jsonrpsee_ws_server::WsServerBuilder;

// Set the ping interval to 10 seconds.
let builder = WsServerBuilder::default().ping_interval(Duration::from_secs(10));

pub fn set_id_provider<I: IdProvider + 'static>(self, id_provider: I) -> Self

Configure custom subscription ID provider for the server to use to when getting new subscription calls.

You may choose static dispatch or dynamic dispatch because IdProvider is implemented for Box<T>.

Default: RandomIntegerIdProvider.

Examples

use jsonrpsee_ws_server::{WsServerBuilder, RandomStringIdProvider, IdProvider};

// static dispatch
let builder1 = WsServerBuilder::default().set_id_provider(RandomStringIdProvider::new(16));

// or dynamic dispatch
let builder2 = WsServerBuilder::default().set_id_provider(Box::new(RandomStringIdProvider::new(16)));

pub fn set_access_control(self, acl: AccessControl) -> Self

Sets access control settings.

pub async fn build(self, addrs: impl ToSocketAddrs) -> Result<Server<M>, Error>

Finalize the configuration of the server. Consumes the Builder.

#[tokio::main]
async fn main() {
  let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
  let occupied_addr = listener.local_addr().unwrap();
  let addrs: &[std::net::SocketAddr] = &[
      occupied_addr,
      "127.0.0.1:0".parse().unwrap(),
  ];
  assert!(jsonrpsee_ws_server::WsServerBuilder::default().build(occupied_addr).await.is_err());
  assert!(jsonrpsee_ws_server::WsServerBuilder::default().build(addrs).await.is_ok());
}