Struct jsonrpsee_ws_server::RpcModule

pub struct RpcModule<Context> { /* fields omitted */ }

Sets of JSON-RPC methods can be organized into a “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

Implementations

impl<Context> RpcModule<Context>

pub fn new(ctx: Context) -> RpcModule<Context>

Create a new module with a given shared Context.

pub fn remove_context(self) -> RpcModule<()>

Transform a module into an RpcModule<()> (unit context).

impl<Context> RpcModule<Context> where
Context: 'static + Send + Sync,

pub fn register_method<R, F>(
&mut self,
method_name: &'static str,
callback: F
) -> Result<MethodResourcesBuilder<'_>, Error> where
Context: 'static + Send + Sync,
R: Serialize,
F: 'static + Fn(Params<'_>, &Context) -> Result<R, Error> + Send + Sync,

Register a new synchronous RPC method, which computes the response with the given callback.

pub fn register_async_method<R, Fun, Fut>(
&mut self,
method_name: &'static str,
callback: Fun
) -> Result<MethodResourcesBuilder<'_>, Error> where
R: 'static + Serialize + Send + Sync,
Fut: Future<Output = Result<R, Error>> + Send,
Fun: 'static + Fn(Params<'static>, Arc<Context>) -> Fut + Copy + Send + Sync,

Register a new asynchronous RPC method, which computes the response with the given callback.

pub fn register_blocking_method<R, F>(
&mut self,
method_name: &'static str,
callback: F
) -> Result<MethodResourcesBuilder<'_>, Error> where
Context: 'static + Send + Sync,
R: Serialize,
F: 'static + Fn(Params<'_>, Arc<Context>) -> Result<R, Error> + Copy + Send + Sync,

Register a new blocking synchronous RPC method, which computes the response with the given callback. Unlike the regular register_method, this method can block its thread and perform expensive computations.

pub fn register_subscription<F>(
&mut self,
subscribe_method_name: &'static str,
notif_method_name: &'static str,
unsubscribe_method_name: &'static str,
callback: F
) -> Result<(), Error> where
Context: 'static + Send + Sync,
F: 'static + Fn(Params<'_>, PendingSubscription, Arc<Context>) + Send + Sync,

Register a new publish/subscribe interface using JSON-RPC notifications.

It implements the ethereum pubsub specification with an option to choose custom subscription ID generation.

Furthermore, it generates the unsubscribe implementation where a bool is used as the result to indicate whether the subscription was successfully unsubscribed to or not. For instance an unsubscribe call may fail if a non-existent subscriptionID is used in the call.

This method ensures that the subscription_method_name and unsubscription_method_name are unique. The notif_method_name argument sets the content of the method field in the JSON document that the server sends back to the client. The uniqueness of this value is not machine checked and it’s up to the user to ensure it is not used in any other RpcModule used in the server.

Arguments

subscription_method_name - name of the method to call to initiate a subscription

notif_method_name - name of method to be used in the subscription payload (technically a JSON-RPC notification)

unsubscription_method - name of the method to call to terminate a subscription

callback - A callback to invoke on each subscription; it takes three parameters:

Params: JSON-RPC parameters in the subscription call.

SubscriptionSink: A sink to send messages to the subscriber.

Context: Any type that can be embedded into the RpcModule.

Examples

use jsonrpsee_core::server::rpc_module::{RpcModule, SubscriptionSink};
use jsonrpsee_core::Error;

let mut ctx = RpcModule::new(99_usize);
ctx.register_subscription("sub", "notif_name", "unsub", |params, pending, ctx| {
    let x = match params.one::<usize>() {
        Ok(x) => x,
        Err(e) => {
            let err: Error = e.into();
            pending.reject(err);
            return;
        }
    };

    let mut sink = match pending.accept() {
        Some(sink) => sink,
        _ => {
           return;
        }
    };

    std::thread::spawn(move || {
        let sum = x + (*ctx);
        let _ = sink.send(&sum);
    });
});

pub fn register_alias(
&mut self,
alias: &'static str,
existing_method: &'static str
) -> Result<(), Error>

Register an alias for an existing_method. Alias uniqueness is enforced.