Struct jsonrpsee_ws_server::WsServerBuilder

pub struct WsServerBuilder<M = ()> { /* fields omitted */ }

Builder to configure and create a JSON-RPC Websocket server

Implementations

impl Builder

pub fn new() -> Self

Create a default server builder.

impl<M> Builder<M>

pub fn max_request_body_size(self, size: u32) -> Self

Set the maximum size of a request body in bytes. Default is 10 MiB.

pub fn max_response_body_size(self, size: u32) -> Self

Set the maximum size of a response body in bytes. Default is 10 MiB.

pub fn max_connections(self, max: u64) -> Self

Set the maximum number of connections allowed. Default is 100.

pub fn batch_requests_supported(self, supported: bool) -> Self

Enables or disables support of batch requests. By default, support is enabled.

pub fn max_subscriptions_per_connection(self, max: u32) -> Self

Set the maximum number of connections allowed. Default is 1024.

pub fn register_resource(
self,
label: &'static str,
capacity: u16,
default: u16
) -> Result<Self, Error>

Register a new resource kind. Errors if label is already registered, or if the number of registered resources on this server instance would exceed 8.

See the module documentation for resurce_limiting for details.

pub fn set_allowed_origins<Origin, List>(
self,
list: List
) -> Result<Self, Error> where
List: IntoIterator<Item = Origin>,
Origin: Into<String>,

Set a list of allowed origins. During the handshake, the Origin header will be checked against the list, connections without a matching origin will be denied. Values should be hostnames with protocol.

builder.set_allowed_origins(["https://example.com"]);

By default allows any Origin.

Will return an error if list is empty. Use allow_all_origins to restore the default.

pub fn set_middleware<T: Middleware>(self, middleware: T) -> Builder<T>

Add a middleware to the builder Middleware.

use std::time::Instant;

use jsonrpsee_core::middleware::Middleware;
use jsonrpsee_ws_server::WsServerBuilder;

#[derive(Clone)]
struct MyMiddleware;

impl Middleware for MyMiddleware {
    type Instant = Instant;

    fn on_request(&self) -> Instant {
        Instant::now()
    }

    fn on_result(&self, name: &str, success: bool, started_at: Instant) {
        println!("Call to '{}' took {:?}", name, started_at.elapsed());
    }
}

let builder = WsServerBuilder::new().set_middleware(MyMiddleware);

pub fn allow_all_origins(self) -> Self

Restores the default behavior of allowing connections with Origin header containing any value. This will undo any list set by set_allowed_origins.

pub fn set_allowed_hosts<Host, List>(self, list: List) -> Result<Self, Error> where
List: IntoIterator<Item = Host>,
Host: Into<String>,

Set a list of allowed hosts. During the handshake, the Host header will be checked against the list. Connections without a matching host will be denied. Values should be hostnames without protocol.

builder.set_allowed_hosts(["example.com"]);

By default allows any Host.

Will return an error if list is empty. Use allow_all_hosts to restore the default.

pub fn allow_all_hosts(self) -> Self

Restores the default behavior of allowing connections with Host header containing any value. This will undo any list set by set_allowed_hosts.

pub fn custom_tokio_runtime(self, rt: Handle) -> Self

Configure a custom tokio::runtime::Handle to run the server on.

Default: tokio::spawn

pub fn set_id_provider<I: IdProvider + 'static>(self, id_provider: I) -> Self

Configure custom subscription ID provider for the server to use to when getting new subscription calls.

You may choose static dispatch or dynamic dispatch because IdProvider is implemented for Box<T>.

Default: RandomIntegerIdProvider.

Examples

use jsonrpsee_ws_server::{WsServerBuilder, RandomStringIdProvider, IdProvider};

// static dispatch
let builder1 = WsServerBuilder::default().set_id_provider(RandomStringIdProvider::new(16));

// or dynamic dispatch
let builder2 = WsServerBuilder::default().set_id_provider(Box::new(RandomStringIdProvider::new(16)));

pub async fn build(self, addrs: impl ToSocketAddrs) -> Result<Server<M>, Error>

Finalize the configuration of the server. Consumes the Builder.

#[tokio::main]
async fn main() {
  let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
  let occupied_addr = listener.local_addr().unwrap();
  let addrs: &[std::net::SocketAddr] = &[
      occupied_addr,
      "127.0.0.1:0".parse().unwrap(),
  ];
  assert!(jsonrpsee_ws_server::WsServerBuilder::default().build(occupied_addr).await.is_err());
  assert!(jsonrpsee_ws_server::WsServerBuilder::default().build(addrs).await.is_ok());
}